contrac-client-gen as a feature used in the public macro
