+++ stylus-proc/Cargo.toml
@@ -37,6 +37,7 @@ default = []
 export-abi = ["stylus-sdk/export-abi"]
 reentrant = []
 stylus-test = ["stylus-sdk/stylus-test"]
+contract-client-gen = []
 
 [package.metadata.docs.rs]
 features = ["export-abi"]
