+++ stylus-proc/src/macros/mod.rs
@@ -1,7 +1,6 @@
 // Copyright 2024, Offchain Labs, Inc.
 // For licensing, see https://github.com/OffchainLabs/stylus-sdk-rs/blob/main/licenses/COPYRIGHT.md
 
-mod contract_client_gen;
 mod derive;
 mod entrypoint;
 mod public;
@@ -9,7 +8,6 @@ mod sol_interface;
 mod sol_storage;
 mod storage;
 
-pub use contract_client_gen::contract_client_gen;
 pub use derive::abi_type::derive_abi_type;
 pub use derive::erase::derive_erase;
 pub use derive::solidity_error::derive_solidity_error;
