+++ stylus-sdk/Cargo.toml
@@ -51,3 +51,4 @@ hostio = []
 mini-alloc = ["dep:mini-alloc"]
 stylus-test = ["dep:stylus-test", "dep:rclite", "stylus-proc/stylus-test"]
 reentrant = ["stylus-proc/reentrant", "stylus-core/reentrant", "stylus-test/reentrant"]
+contract-client-gen = ["stylus-proc/contract-client-gen"]
